auto size give graph
